Move package files out of `src`
