+++ packages/bento-common/utils/Base64.ts
