+++ packages/bento-core/tokens/bnb.json
