+++ packages/bento-core/tokens/ethereum.json
