+++ packages/bento-core/tokens/klaytn.json
