+++ packages/bento-core/tokens/solana.json
