+++ packages/bento-web/src/dashboard/utils/useNFTBalances.ts
@@ -1,6 +1,6 @@
 import { OpenSea, OpenSeaAsset } from '@bento/client';
 import { Wallet, safePromiseAll } from '@bento/common';
-import { priceFromCoinGecko } from '@bento/core/lib/pricings/CoinGecko';
+import { priceFromCoinGecko } from '@bento/core/pricings/CoinGecko';
 import chunk from 'lodash.chunk';
 import groupBy from 'lodash.groupby';
 import { useEffect, useState } from 'react';
