+++ packages/bento-web/src/pages/api/cosmos-sdk/[network]/[walletAddress].ts
@@ -1,12 +1,12 @@
 import { CosmosSDKBasedNetworks } from '@bento/common';
 import { safePromiseAll } from '@bento/common';
-import { Bech32Address } from '@bento/core/lib/bech32';
+import { Bech32Address } from '@bento/core/bech32';
 import {
   CosmosHubChain,
   CosmosSDKBasedChain,
   OsmosisChain,
   TokenBalance,
-} from '@bento/core/lib/chains';
+} from '@bento/core/chains';
 import type { NextApiRequest, NextApiResponse } from 'next';
 
 interface APIRequest extends NextApiRequest {
